Uses satoshi field formatter on custom amount dialog
