+++ lib/widgets/amount_form_field.dart
@@ -2,6 +2,7 @@ import 'package:breez/bloc/account/account_model.dart';
 import 'package:breez/bloc/user_profile/currency.dart';
 import 'package:breez/theme_data.dart' as theme;
 import 'package:breez/widgets/currency_converter_dialog.dart';
+import 'package:breez/widgets/sat_amount_form_field_formatter.dart';
 import 'package:fixnum/fixnum.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -101,53 +102,5 @@ class AmountFormField extends TextFormField {
         return "Invalid amount";
       }
     };
-  }
-}
-
-class SatAmountFormFieldFormatter extends TextInputFormatter {
-  final RegExp _pattern = RegExp(r'[^\d*]');
-
-  TextEditingValue formatEditUpdate(
-    TextEditingValue oldValue,
-    TextEditingValue newValue,
-  ) {
-    final raw = newValue.text.replaceAll(_pattern, '');
-    if (raw.isEmpty) {
-      return newValue.copyWith(
-        text: '',
-        selection: TextSelection.collapsed(offset: 0),
-      );
-    }
-
-    var value;
-    try {
-      value = Int64.parseInt(raw.length > 18 ? raw.substring(0, 18) : raw);
-    } catch (ignored) {
-      value = Int64(0);
-    }
-
-    final formatted = Currency.SAT.format(
-      value,
-      includeDisplayName: false,
-      includeCurrencySymbol: false,
-    );
-
-    var diff = formatted.length - oldValue.text.length;
-    var newOffset = newValue.selection.start;
-    if (formatted != oldValue.text) {
-      if (diff > 1) {
-        newOffset += 1;
-      }
-      if (diff < -1) {
-        newOffset -= 1;
-      }
-    } else {
-      newOffset = oldValue.selection.start;
-    }
-
-    return newValue.copyWith(
-      text: formatted,
-      selection: TextSelection.collapsed(offset: newOffset),
-    );
   }
 }
